Partial reversal of e062cc4b6e5b43201f62713bd885d2951077b869
Currently the Gmail IMAP server is not returning the \Inbox label as a
X-GM-LABELS for a message, and remove \Inbox from X-GM-LABELS does not
remove the email from the inbox. This is possibly a bug with Gmail's
IMAP API, the reference isn't clear on the expected behavior.

Using the standard IMAP copy and delete method for all servers, removing
the special case for servers which claim support for X-GM-EXT-1 (aka
Gmail).